Change thread shutdown to best-effort on Unix
Shutting down a Unix thread is mega-hard and apparently not very portable. Plus
it's basically racy as we're simply pausing for a bit waiting for a thread to
shut down, but in theory could take awhile.

Observed in https://bugzilla.mozilla.org/show_bug.cgi?id=1445413 it's not really
critical to unconditionally always shut down the helper thread and otherwise
it's generally a bad idea to panic in destructors.

As a result basically just ignore the ability to shut down the helper thread.