Rewrite with Mutex/Condvar intead of `mpsc`
The `std::sync::mpsc` module is quite heavy in terms of codegen, so make
this crate a bit of a lighter dependency by using mutexes/condvars which
are much easier on codegen.

On a local machine in `-j1` mode this goes from 1.97s to 1.28s in
release mode and 1.61s to 1.11s in debug mode.