Handle EINTR.
When building with sccache locally, I'm seeing a bunch of errors to acquire the
jobserver token which are caused by interrupts at bad times.

I think it was a kernel update what caused this, fwiw, but not 100% sure.

Both can happen according to the man pages, and we're not handling them
sensibly.

I think retrying the read / poll when the error is EINTR is a sensible thing to
do.